+++ src/common/enums.ts
@@ -1,3 +1,28 @@
+export enum Glyph {
+  Implication = '=>',
+  Conjunction = '&',
+  Disjunction = '|',
+  Negation = '~',
+  Equivalence = '<=>',
+  OpenParenthesis = '(',
+  CloseParenthesis = ')',
+}
+
+export enum GlyphType {
+  Variable = 'variable',
+  Operator = 'operator',
+  Parenthesis = 'parenthesis',
+}
+
+export enum Operator {
+  Var = 'VAR',
+  Not = 'NOT',
+  Or = 'OR',
+  And = 'AND',
+  Implies = 'IMPLIES',
+  Equiv = 'EQUIV',
+}
+
 export enum StepReason {
   Premise = 'Premise',
   Assumption = 'Assumption',
