+++ src/common/types.ts
@@ -1,5 +1,20 @@
-import { PropExpression, PropFormula } from './base';
-import { StepReason } from '../../enums';
+import { GlyphType, Operator, StepReason } from './enums';
+
+export type PropAtom = [string];
+
+export type PropSymbol = {
+  atom: PropAtom;
+  type: GlyphType;
+  position: number;
+  view: string;
+};
+
+export type PropExpression = PropSymbol[];
+
+export type PropFormula = {
+  operator: Operator;
+  values: PropFormula[] | PropAtom;
+};
 
 export type ProofStep = {
   id: string;
