--- src/enums/glyph-type.ts
@@ -1,5 +1,0 @@
-export enum GlyphType {
-  Variable = 'variable',
-  Operator = 'operator',
-  Parenthesis = 'parenthesis',
-}
