--- src/enums/glyph.ts
@@ -1,9 +1,0 @@
-export enum Glyph {
-  Implication = '=>',
-  Conjunction = '&',
-  Disjunction = '|',
-  Negation = '~',
-  Equivalence = '<=>',
-  OpenParenthesis = '(',
-  CloseParenthesis = ')',
-}
