--- src/enums/index.ts
@@ -1,4 +1,0 @@
-export { GlyphType } from './glyph-type';
-export { Operator } from './operator';
-export { StepReason } from './step-reason';
-export { Glyph } from './glyph';
