--- src/enums/operator.ts
@@ -1,8 +1,0 @@
-export enum Operator {
-  Var = 'VAR',
-  Not = 'NOT',
-  Or = 'OR',
-  And = 'AND',
-  Implies = 'IMPLIES',
-  Equiv = 'EQUIV',
-}
