--- src/models/index.ts
@@ -1,2 +1,0 @@
-export type { PropAtom, PropSymbol, PropExpression, PropFormula } from './propositions/base.ts';
-export type { ProofStep, NaturalProofStep, DirectProofStep } from './propositions/proofs.ts';
